backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pandas
openpyxl
xlrd>=2.0.1
pdfplumber
psycopg2-binary
python-dotenv
